requirements.txt
aiohttp==3.7.4.post0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
braceexpand==0.1.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
cytoolz==0.11.2
discord-py-interactions==3.0.2
discord.py==1.7.3
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
h11==0.13.0
hexbytes==0.2.2
idna==3.3
import-expression==1.1.4
ipfshttpclient==0.8.0a2
jishaku==2.3.2
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
outcome==1.1.0
parsimonious==0.8.1
Pillow==9.0.1
protobuf==3.19.4
pycparser==2.21
pycryptodome==3.14.1
pyOpenSSL==22.0.0
pyrsistent==0.18.1
PySocks==1.7.1
python-dotenv==0.19.2
requests==2.27.1
rlp==2.0.1
selenium==4.1.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toolz==0.11.2
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.1.1
uniswap-python==0.5.5
urllib3==1.26.8
varint==1.0.2
web3==5.28.0
websockets==9.1
wsproto==1.1.0
yarl==1.7.2
